+++ CHANGES.md
@@ -10,6 +10,7 @@
  * [METAMODEL-215] - Improved the capability of NumberComparator to support Integer, Long, Double, BigInteger and other built-in Number classes.
  * [METAMODEL-218] - Fixed conversion of STRING and NUMBER types to database-specific types in JDBC module.
  * [METAMODEL-205] - Added validation of Excel sheet name before attempting to create table (sheet).
+ * [METAMODEL-219] - Made HdfsResource capable of incorporating Hadoop configuration files core-site.xml and hdfs-site.xml
 
 ### Apache MetaModel 4.4.1